requirements.txt
torch>=2.0.0
torch-geometric>=2.4.0
rdkit>=2023.3.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.31.0
streamlit>=1.28.0
plotly>=5.17.0
Pillow>=10.0.0
networkx>=3.1
imbalanced-learn>=0.11.0
